requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
deprecated==1.2.18
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.60.2
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pygithub==2.5.0
pyjwt==2.10.1
pynacl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
